+++ src/Browser/State.php
@@ -70,6 +70,5 @@ class State
         }
 
         return $this->crawler;
-
     }
 }